requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.0
beautifulsoup4==4.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
compressed-rtf==1.0.7
comtypes==1.4.10; platform_system == "Windows"
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
ebcdic==1.1.1
et_xmlfile==2.0.0
eventlet==0.39.1
exceptiongroup==1.2.2
extract-msg==0.23.1
fastapi==0.115.12
Flask==2.3.3
Flask-SocketIO==5.5.1
google==3.0.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
MarkupSafe==3.0.2
numpy==2.0.2
olefile==0.46
openai==1.66.3
openpyxl==3.1.2
pdfminer.six==20181108
pillow==11.1.0
protobuf==6.30.2
py==1.11.0
pycryptodome==3.9.9
pydantic==2.11.3
pydantic_core==2.33.1
PyPDF2==3.0.1
pypiwin32==223; platform_system == "Windows"
python-docx==0.8.11
python-engineio==4.12.0
python-pptx==0.6.18
python-socketio==5.13.0
pyttsx3==2.98
pytz==2020.5
requests==2.31.0
retry==0.9.2
simple-websocket==1.1.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
starlette==0.46.2
textract==1.6.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==1.5.1
urllib3==2.4.0
uvicorn==0.34.1
volcengine==1.0.180
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.2
zipp==3.21.0
python-dotenv==1.0.1
flask-restx==1.3.0

# 测试依赖
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
